+++ packages/vrender-components/src/label/arc.ts
@@ -511,7 +511,8 @@ export class ArcLabel extends LabelBase<ArcLabelAttrs> {
     const flag = isQuadrantLeft(quadrant) ? -1 : 1;
     let cx: number = 0;
     let limit =
-      (flag > 0 ? plotLayout.x2 - pointB.x + this._alignOffset : pointB.x - plotLayout.x1 - this._alignOffset) -
+      (flag > 0 ? plotLayout.x2 - pointB.x : pointB.x - plotLayout.x1) +
+      this._alignOffset -
       this._line2MinLength -
       spaceWidth;