+++ packages/vrender-components/src/legend/discrete/discrete.ts
@@ -626,10 +626,10 @@ export class DiscreteLegend extends LegendBase<DiscreteLegendAttrs> {
             x: this._itemWidthByUser - shapeSize / 2 - parsedPadding[1] - parsedPadding[3] - focusSpace - valueSpace
           });
         } else {
-          valueShape.setAttribute('x', valueSpace + (labelShape.AABBBounds.empty() ? 0 : labelShape.AABBBounds.x2));
+          valueShape.setAttribute('x', labelSpace + (labelShape.AABBBounds.empty() ? 0 : labelShape.AABBBounds.x2));
         }
       } else {
-        valueShape.setAttribute('x', valueSpace + (labelShape.AABBBounds.empty() ? 0 : labelShape.AABBBounds.x2));
+        valueShape.setAttribute('x', labelSpace + (labelShape.AABBBounds.empty() ? 0 : labelShape.AABBBounds.x2));
       }
       focusStartX = valueSpace + (valueShape.AABBBounds.empty() ? 0 : valueShape.AABBBounds.x2);
 
@@ -740,14 +740,8 @@ export class DiscreteLegend extends LegendBase<DiscreteLegendAttrs> {
         });
   }
 
-  private _updatePositionOfPager(
-    contentWidth: number,
-    contentHeight: number,
-    renderStartY: number,
-    compWidth: number,
-    compHeight: number
-  ) {
-    const { maxHeight, pager } = this.attribute;
+  private _updatePositionOfPager(renderStartY: number, compWidth: number, compHeight: number) {
+    const { pager } = this.attribute;
     const { totalPage, isHorizontal } = this._itemContext;
     const position = (pager && (pager as LegendPagerAttributes).position) || 'middle';
     (this._pagerComponent as Pager).setTotal(totalPage);
@@ -762,7 +756,7 @@ export class DiscreteLegend extends LegendBase<DiscreteLegendAttrs> {
         y = renderStartY + compHeight / 2 - this._pagerComponent.AABBBounds.height() / 2;
       }
       this._pagerComponent.setAttributes({
-        x: contentWidth,
+        x: compWidth - this._pagerComponent.AABBBounds.width(),
         y
       });
     } else {
@@ -776,7 +770,7 @@ export class DiscreteLegend extends LegendBase<DiscreteLegendAttrs> {
       }
       this._pagerComponent.setAttributes({
         x,
-        y: (maxHeight as number) - this._pagerComponent.AABBBounds.height()
+        y: compHeight - this._pagerComponent.AABBBounds.height()
       });
     }
   }
@@ -805,12 +799,24 @@ export class DiscreteLegend extends LegendBase<DiscreteLegendAttrs> {
     const pageParser = this._itemContext.isScrollbar
       ? (e: CustomEvent) => {
           const { value } = e.detail;
-          let newPage = value[0] * this._itemContext.totalPage;
-
-          if ((pager as LegendScrollbarAttributes).scrollByPosition) {
-            newPage = newPage + 1;
+          let newPage;
+          // 判断scrollbar在起点和终点的临界情况
+          // 原因:
+          // 假设totalPage为6, 则初始range=[0, 0.16666666666666666], 此时span=0.16666666666666666 = 1/6
+          // 随着scrollbar移动, 计算精度有所偏差, 在终点时range=[0.833333333333333, 1], 此时span=0.16666666666666696 > 1/6
+          // 导致在终点时无法触及最后一页的起点, 从而无法成功滚动到最后一页
+          if (value[0] === 0) {
+            newPage = 1;
+          } else if (value[1] === 1) {
+            newPage = this._itemContext.totalPage;
           } else {
-            newPage = Math.floor(newPage) + 1;
+            newPage = value[0] * this._itemContext.totalPage;
+
+            if ((pager as LegendScrollbarAttributes).scrollByPosition) {
+              newPage = newPage + 1;
+            } else {
+              newPage = Math.floor(newPage) + 1;
+            }
           }
 
           return newPage;
@@ -820,7 +826,6 @@ export class DiscreteLegend extends LegendBase<DiscreteLegendAttrs> {
         };
 
     const onScroll = (e: FederatedWheelEvent) => {
-      e.preventDefault();
       const scrollComponent = this._pagerComponent as ScrollBar;
       const preScrollRange = scrollComponent.getScrollRange();
       const { direction } = scrollComponent.attribute as ScrollBarAttributes;
@@ -863,6 +868,8 @@ export class DiscreteLegend extends LegendBase<DiscreteLegendAttrs> {
       this._pagerComponent.addEventListener('scrollUp', onPaging);
       if (((this.attribute as DiscreteLegendAttrs).pager as LegendScrollbarAttributes).roamScroll) {
         const THROTTLE_TIME = 50;
+        // preventDefault不能和throttle一起使用, 否则阻止默认事件失败
+        this.addEventListener('wheel', (e: FederatedWheelEvent) => e.nativeEvent.preventDefault());
         this.addEventListener('wheel', throttle(onScroll, THROTTLE_TIME));
       }
     } else {
@@ -928,7 +935,7 @@ export class DiscreteLegend extends LegendBase<DiscreteLegendAttrs> {
 
       this._itemContext.totalPage = total;
 
-      this._updatePositionOfPager(contentWidth, contentHeight, renderStartY, compWidth, compHeight);
+      this._updatePositionOfPager(renderStartY, compWidth, compHeight);
     } else {
       compWidth = this._itemMaxWidth * maxCol + (maxCol - 1) * spaceCol;
       compHeight = maxHeight;
@@ -969,7 +976,7 @@ export class DiscreteLegend extends LegendBase<DiscreteLegendAttrs> {
       const total = Math.ceil(pages / maxCol);
 
       this._itemContext.totalPage = total;
-      this._updatePositionOfPager(contentWidth, contentHeight, renderStartY, compWidth, compHeight);
+      this._updatePositionOfPager(renderStartY, compWidth, compHeight);
     }
 
     // 初始化 defaultCurrent
@@ -1376,5 +1383,10 @@ export class DiscreteLegend extends LegendBase<DiscreteLegendAttrs> {
     }
 
     return newConfig;
+  }
+
+  release(): void {
+    super.release();
+    this.removeAllEventListeners();
   }
 }