+++ packages/vrender-core/src/graphic/config.ts
@@ -125,6 +125,8 @@ export const DefaultStyle: IGraphicStyle = {
   forceBoundsHeight: undefined,
   opacity: 1,
   background: null,
+  textureRatio: 0,
+  textureOptions: null,
   backgroundOpacity: 1,
   backgroundCornerRadius: 0,
   texture: null,
@@ -177,6 +179,7 @@ export const DefaultAttribute: Required<IGraphicAttribute> = {
   zIndex: 0,
   layout: null,
   boundsPadding: 0,
+  fillStrokeOrder: 0,
   renderStyle: 'default',
   pickMode: 'accurate',
   customPickShape: null,
@@ -275,6 +278,8 @@ export const DefaultLineAttribute: Required<ILineGraphicAttribute> = {
 export const DefaultPathAttribute: Required<IPathGraphicAttribute> = {
   ...DefaultAttribute,
   path: new CustomPath2D(),
+  fillStrokeOrder: 1,
+  clipRange: 1,
   customPath: () => {
     Logger.getInstance().warn('空函数');
   }
@@ -311,7 +316,8 @@ export const DefaultSymbolAttribute: Required<ISymbolGraphicAttribute> = {
   ...DefaultAttribute,
   symbolType: 'circle',
   size: 10, // 外接**正方形**的边长
-  keepDirIn3d: true
+  keepDirIn3d: true,
+  clipRange: 1
 };
 
 export const DefaultTextAttribute: Required<ITextGraphicAttribute> = {