+++ packages/vrender-core/src/render/contributions/render/polygon-render.ts
@@ -16,11 +16,9 @@ import type {
   IRenderService
 } from '../../../interface';
 import { drawPolygon, drawRoundedPolygon } from '../../../common/polygon';
-import { drawPathProxy, fillVisible, runFill, runStroke, strokeVisible } from './utils';
 import { PolygonRenderContribution } from './contributions/constants';
 // eslint-disable-next-line @typescript-eslint/consistent-type-imports
 import { ContributionProvider } from '../../../common/contribution-provider';
-import { BaseRenderContributionTime } from '../../../common/enums';
 import { BaseRender } from './base-render';
 import {
   defaultPolygonBackgroundRenderContribution,
@@ -67,7 +65,8 @@ export class DefaultCanvasPolygonRender extends BaseRender<IPolygon> implements 
       cornerRadius = polygonAttribute.cornerRadius,
       x: originX = polygonAttribute.x,
       y: originY = polygonAttribute.y,
-      closePath = polygonAttribute.closePath
+      closePath = polygonAttribute.closePath,
+      fillStrokeOrder = polygonAttribute.fillStrokeOrder
     } = polygon.attribute;
 
     const data = this.valid(polygon, polygonAttribute, fillCb, strokeCb);
@@ -105,23 +104,35 @@ export class DefaultCanvasPolygonRender extends BaseRender<IPolygon> implements 
       strokeCb
     );
 
-    if (doFill) {
-      if (fillCb) {
-        fillCb(context, polygon.attribute, polygonAttribute);
-      } else if (fVisible) {
-        // 存在fill
-        context.setCommonStyle(polygon, polygon.attribute, originX - x, originY - y, polygonAttribute);
-        context.fill();
+    const _runFill = () => {
+      if (doFill) {
+        if (fillCb) {
+          fillCb(context, polygon.attribute, polygonAttribute);
+        } else if (fVisible) {
+          // 存在fill
+          context.setCommonStyle(polygon, polygon.attribute, originX - x, originY - y, polygonAttribute);
+          context.fill();
+        }
       }
-    }
-    if (doStroke) {
-      if (strokeCb) {
-        strokeCb(context, polygon.attribute, polygonAttribute);
-      } else if (sVisible) {
-        // 存在stroke
-        context.setStrokeStyle(polygon, polygon.attribute, originX - x, originY - y, polygonAttribute);
-        context.stroke();
+    };
+    const _runStroke = () => {
+      if (doStroke) {
+        if (strokeCb) {
+          strokeCb(context, polygon.attribute, polygonAttribute);
+        } else if (sVisible) {
+          // 存在stroke
+          context.setStrokeStyle(polygon, polygon.attribute, originX - x, originY - y, polygonAttribute);
+          context.stroke();
+        }
       }
+    };
+
+    if (!fillStrokeOrder) {
+      _runFill();
+      _runStroke();
+    } else {
+      _runStroke();
+      _runFill();
     }
 
     this.afterRenderStep(