+++ packages/vrender/__tests__/browser/src/pages/text.ts
@@ -217,19 +217,19 @@ export const page = () => {
   graphics.push(rect);
 
   const textLimit = createText({
-    x: 800,
-    y: 500,
+    x: 500,
+    y: 300,
     fill: colorPools[5],
     // text: ['Tffg'],
-    text: '这是textabc这aaaaa是什么这是阿萨姆abcaaaaabcdef这是textabc这aaaaa是什么这是阿萨姆abcaaaaa',
+    text: 'this is textaaaaaaaaaaaaaaaaa aaa this isisisisisis abc',
     // text: '这是textabc这aaaaa是什么这是阿萨姆abcaaaaabcdef这是textabc这aaaaa是什么这是阿萨姆abcaaaaa',
     // heightLimit: 40,
-    wordBreak: 'break-word',
-    maxLineWidth: 200,
+    wordBreak: 'keep-all',
+    maxLineWidth: 100,
     stroke: 'green',
     textAlign: 'left',
-    textBaseline: 'middle'
-    // whiteSpace: 'normal'
+    textBaseline: 'middle',
+    whiteSpace: 'normal'
     // wrap: true
   });
   console.log('textLimit', textLimit);