+++ packages/react-vrender-utils/CHANGELOG.md
@@ -1,6 +1,16 @@
 # Change Log - @visactor/react-vrender-utils
 
-This log was last generated on Sun, 29 Sep 2024 09:44:02 GMT and should not be manually modified.
+This log was last generated on Wed, 23 Oct 2024 08:37:33 GMT and should not be manually modified.
+
+## 0.20.10
+Wed, 23 Oct 2024 08:37:33 GMT
+
+_Version update only_
+
+## 0.20.9
+Tue, 15 Oct 2024 03:50:15 GMT
+
+_Version update only_
 
 ## 0.20.8
 Sun, 29 Sep 2024 09:44:02 GMT